requirements.txt
fastapi
uvicorn

# settings
python-dotenv
pydantic-settings

# database
sqlalchemy
asyncpg
psycopg2-binary
alembic

# jwt
pyjwt
bcrypt

jinja2
